+++ src/ng-select/public-api.ts
@@ -2,7 +2,7 @@
  * Public API Surface of ng-select
  */
 
-export { NgSelectComponent, SELECTION_MODEL_FACTORY } from './lib/ng-select.component';
+export { DropdownPosition, NgSelectComponent, SELECTION_MODEL_FACTORY } from './lib/ng-select.component';
 export { NgSelectModule } from './lib/ng-select.module';
 export { NgOption } from './lib/ng-select.types';
 export { SelectionModel } from './lib/selection-model';